+++ NEWS.md
@@ -1,5 +1,7 @@
 # mirai (development version)
 
+* Requires nanonext >= [1.7.2.9000].
+
 # mirai 2.5.3
 
 #### Updates