+++ NEWS.md
@@ -1,5 +1,8 @@
 # mirai (development version)
 
+#### Updates
+
+* A `miraiError` gains `conditionCall` and `conditionMessage` methods (@jan-swissre, #529) 
 * Requires nanonext >= [1.7.2.9000].
 
 # mirai 2.5.3