+++ NEWS.md
@@ -8,6 +8,7 @@
 
 #### Updates
 
+* Fixes an issue in `daemons()` where supplying a non-character value to `url` causes local daemons to be launched instead of thowing an error (#549).
 * Fixes a bug whereby repeated mirai cancellation could sometimes cause a daemon to exit prematurely (#546).
 * A `miraiError` gains `conditionCall` and `conditionMessage` methods (@jan-swissre, #529)
 * Requires nanonext >= [1.7.2.9000].