+++ NEWS.md
@@ -13,7 +13,7 @@
 #### Updates
 
 * The `dispatcher()` loop has been re-implemented entirely in C code in nanonext, and now has roughly half the previous overhead (#527).
-* Fixes an issue in `daemons()` where supplying a non-character value to `url` causes local daemons to be launched instead of thowing an error (#549).
+* Fixes an issue in `daemons()` where supplying a non-character value to `url` causes local daemons to be launched instead of throwing an error (#549).
 * Fixes a bug whereby repeated mirai cancellation could sometimes cause a daemon to exit prematurely (#546).
 * A `miraiError` gains `conditionCall` and `conditionMessage` methods (@jan-swissre, #529)
 * Requires nanonext >= 1.8.0.